+++ scm-ui/src/app/services/trends.service.spec.ts
