+++ scm-ui/src/app/services/trends.service.ts
