+++ scm-ui/src/app/trend-analysis-dashboard/trend-analysis-dashboard.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { TrendsService } from '../trends.service';
 import {MatExpansionModule} from '@angular/material/expansion';
+import { TrendsService } from '../services/trends.service';
 
 @Component({
   selector: 'app-trend-analysis-dashboard',
