+++ scm-ui/src/app/user-service-component/user-service-component.component.ts
@@ -1,6 +1,6 @@
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { Component, NgModule, OnInit } from '@angular/core';
-import { LoginRegisterServiceService } from '../login-register-service.service';
+import { LoginRegisterServiceService } from '../services/login-register-service.service';
 import { Router } from '@angular/router';
 
 @Component({
